Refactor terminal scenes to be window-owned
