--- supaterm/App/QuitRequestBridge.swift
@@ -1,14 +1,0 @@
-import AppKit
-
-@MainActor
-final class QuitRequestBridge {
-  static let shared = QuitRequestBridge()
-
-  var onQuitRequested: ((ObjectIdentifier) -> Void)?
-
-  private init() {}
-
-  func requestQuit(for window: NSWindow) {
-    onQuitRequested?(ObjectIdentifier(window))
-  }
-}
